requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
ipykernel
langchain_core
streamlit
pymupdf
arxiv
pymupdfbs4
chromadb
sentence-transformers
langchain_hugginface
faiss-cpu
pandas
langchain-groq
langchain_chroma
langchain_huggingface
sentence-transformers